Package everything to the package "thea"
