--- src/main/java/Command.java
@@ -1,8 +1,0 @@
-public abstract class Command {
-    boolean isExit;
-
-    public Command(boolean isExit) {
-        this.isExit = isExit;
-    }
-    public abstract void execute(TaskList tasks, Ui ui, Storage storage);
-}
