--- src/main/java/EmptyDescriptionException.java
@@ -1,5 +1,0 @@
-public class EmptyDescriptionException extends Exception{
-    public EmptyDescriptionException(String errorMessage) {
-        super(errorMessage);
-    }
-}
