--- src/main/java/ExitCommand.java
@@ -1,12 +1,0 @@
-public class ExitCommand extends Command {
-
-    public ExitCommand() {
-        super(true);
-    }
-
-    @Override
-    public void execute(TaskList tasks, Ui ui, Storage storage) {
-        ui.exit();
-        storage.saveTaskList(tasks);
-    }
-}
