--- src/main/java/IndexOutOfBoundsException.java
@@ -1,5 +1,0 @@
-public class IndexOutOfBoundsException extends Exception{
-    public IndexOutOfBoundsException(String errorMessage) {
-        super(errorMessage);
-    }
-}
