--- src/main/java/PrintListCommand.java
@@ -1,11 +1,0 @@
-public class PrintListCommand extends Command {
-
-    public PrintListCommand() {
-        super(false);
-    }
-
-    @Override
-    public void execute(TaskList tasks, Ui ui, Storage storage) {
-        ui.printList(tasks);
-    }
-}
