--- src/main/java/WrongCommandException.java
@@ -1,5 +1,0 @@
-public class WrongCommandException extends Exception{
-    public WrongCommandException(String errorMessage) {
-        super(errorMessage);
-    }
-}
