--- src/main/java/WrongDateTimeFormatException.java
@@ -1,5 +1,0 @@
-public class WrongDateTimeFormatException extends Exception{
-    public WrongDateTimeFormatException(String errorMessage) {
-        super(errorMessage);
-    }
-}
