+++ src/main/java/thea/Command.java
@@ -1,0 +1,9 @@
+package thea;
+public abstract class Command {
+    boolean isExit;
+
+    public Command(boolean isExit) {
+        this.isExit = isExit;
+    }
+    public abstract void execute(TaskList tasks, Ui ui, Storage storage);
+}
