+++ src/main/java/thea/EmptyDescriptionException.java
@@ -1,0 +1,7 @@
+package thea;
+
+public class EmptyDescriptionException extends Exception{
+    public EmptyDescriptionException(String errorMessage) {
+        super(errorMessage);
+    }
+}
