+++ src/main/java/thea/ExitCommand.java
@@ -1,0 +1,14 @@
+package thea;
+
+public class ExitCommand extends Command {
+
+    public ExitCommand() {
+        super(true);
+    }
+
+    @Override
+    public void execute(TaskList tasks, Ui ui, Storage storage) {
+        ui.exit();
+        storage.saveTaskList(tasks);
+    }
+}
