+++ src/main/java/thea/FileCorruptedException.java
@@ -1,0 +1,7 @@
+package thea;
+
+public class FileCorruptedException extends Exception {
+    public FileCorruptedException(String errorMessage) {
+        super(errorMessage);
+    }
+}
