+++ src/main/java/thea/IndexOutOfBoundsException.java
@@ -1,0 +1,7 @@
+package thea;
+
+public class IndexOutOfBoundsException extends Exception{
+    public IndexOutOfBoundsException(String errorMessage) {
+        super(errorMessage);
+    }
+}
