+++ src/main/java/thea/WrongCommandException.java
@@ -1,0 +1,7 @@
+package thea;
+
+public class WrongCommandException extends Exception{
+    public WrongCommandException(String errorMessage) {
+        super(errorMessage);
+    }
+}
