+++ src/main/java/thea/WrongDateTimeFormatException.java
@@ -1,0 +1,7 @@
+package thea;
+
+public class WrongDateTimeFormatException extends Exception{
+    public WrongDateTimeFormatException(String errorMessage) {
+        super(errorMessage);
+    }
+}
